Fix problem with code reloads in spree-backend
The InMemoryStore is a singleton and store Spree preferences in memory,
in that singleton. Every time Ruby would reload the code (when you'd
change something in spree-backend) it would re-load the InMemoryStore
class and re-initialize that singleton. That means you'd lose all Spree
settings, including require_authentication for the API. We set it to
false, but the default is true. And boom you'd get the 401.

This reloads all the settings before every request in development, and
only once in production. That way the settings, including
require_authentication, get re-initialized correctly every time.

@richardnuno this should solve the image problem you were seeing, too,
I'd imagine.